Move author and category back to zinnia.models
